Fix MSVC warnings in jsscope.cpp
UltraBlame original commit: 8cdcb8f8aadbee7d4d2eea47174d3a78c4563f5a

+++ js/src/jsscope.cpp
@@ -973,9 +973,9 @@ JSObject::putProperty(JSContext *cx, jsid id,
 
         shape->rawGetter = getter;
         shape->rawSetter = setter;
-        shape->attrs = attrs;
+        shape->attrs = uint8(attrs);
         shape->flags = flags | Shape::IN_DICTIONARY;
-        shape->shortid = shortid;
+        shape->shortid = int16(shortid);
